requirements.txt
cloudscraper
beautifulsoup4
Pillow
send2trash
requests